moved files to adapters folder
